backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pydantic
python-dotenv
pydantic-settings
